feat(searchbox): create searchbox component
